padme-monitoring/backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
requires-python = ">=3.11,<4.0"
dependencies = [
  "fastapi[standard]<1.0.0,>=0.115.5",
  "pydantic>2.0",
  "psycopg2<3.0.0,>=2.9.10",
  "psycopg2-binary<3.0.0,>=2.9.10",
  "pydantic-settings<3.0.0,>=2.7.0",
  "python-keycloak<6.0.0,>=5.1.0",
  "rdflib<8.0.0,>=7.1.1",
  "rdflib-sqlalchemy<1.0.0,>=0.5.4",
  "sse-starlette<3.0.0,>=2.2.1",
  "setuptools<76.0.0,>=75.1.0",
  "redis>=5.2.1",
]

[tool.uv]
dev-dependencies = [
  "pytest<8.0.0,>=7.4.3",
  "mypy<2.0.0,>=1.8.0",
  "ruff<1.0.0,>=0.2.2",
  "pre-commit<4.0.0,>=3.6.2",
  "types-passlib<2.0.0.0,>=1.7.7.20240106",
  "coverage<8.0.0,>=7.4.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py310"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
  "E",      # pycodestyle errors
  "W",      # pycodestyle warnings
  "F",      # pyflakes
  "I",      # isort
  "B",      # flake8-bugbear
  "C4",     # flake8-comprehensions
  "UP",     # pyupgrade
  "ARG001", # unused arguments in functions
]
ignore = [
  "E501", # line too long, handled by black
  "B008", # do not perform function calls in argument defaults
  "W191", # indentation contains tabs
  "B904", # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
